Consolidate utilities in standard .util; warn on import from .utils
